+++ kwimage/transform.py
@@ -7,6 +7,13 @@ import kwarray
 import skimage.transform
 
 
+try:
+    import xdev
+    profile = xdev.profile
+except Exception:
+    profile = ub.identity
+
+
 class Transform(ub.NiceRepr):
     pass
 
@@ -136,6 +143,7 @@ class Matrix(Transform):
         else:
             raise TypeError('{} @ {}'.format(type(self), type(other)))
 
+    @profile
     def inv(self):
         """
         Returns the inverse of this matrix
@@ -300,6 +308,7 @@ class Affine(Projective):
         return params
 
     @classmethod
+    @profile
     def coerce(cls, data=None, **kwargs):
         """
         Attempt to coerce the data into an affine object
@@ -335,15 +344,16 @@ class Affine(Projective):
             self = data
         elif isinstance(data, dict):
             keys = set(data.keys())
-            known_params = {'scale', 'shear', 'offset', 'theta', 'type'}
-            params = ub.dict_isect(data, known_params)
             if 'matrix' in keys:
                 self = cls(matrix=np.array(data['matrix']))
-            elif len(known_params & keys):
-                params.pop('type', None)
-                self = cls.affine(**params)
             else:
-                raise KeyError(', '.join(list(data.keys())))
+                known_params = {'scale', 'shear', 'offset', 'theta', 'type'}
+                params = {key: data[key] for key in known_params if key in data}
+                if len(known_params & keys):
+                    params.pop('type', None)
+                    self = cls.affine(**params)
+                else:
+                    raise KeyError(', '.join(list(data.keys())))
         else:
             raise TypeError(type(data))
         return self
@@ -404,7 +414,15 @@ class Affine(Projective):
         Returns:
             Affine
         """
-        return cls.affine(scale=scale)
+        scale_ = 1 if scale is None else scale
+        sx, sy = _ensure_iterable2(scale_)
+        # Sympy simplified expression
+        mat = np.array([sx , 0.0, 0.0,
+                        0.0,  sy, 0.0,
+                        0.0, 0.0, 1.0])
+        mat = mat.reshape(3, 3)  # Faster to make a flat array and reshape
+        self = cls(mat)
+        return self
 
     @classmethod
     def translate(cls, offset):
@@ -416,8 +434,33 @@ class Affine(Projective):
 
         Returns:
             Affine
+
+        Benchmark:
+            >>> # xdoctest: +REQUIRES(--benchmark)
+            >>> # It is ~3x faster to use the more specific method
+            >>> import timerit
+            >>> import kwimage
+            >>> #
+            >>> offset = np.random.rand(2)
+            >>> ti = timerit.Timerit(100, bestof=10, verbose=2)
+            >>> for timer in ti.reset('time'):
+            >>>     with timer:
+            >>>         kwimage.Affine.translate(offset)
+            >>> #
+            >>> for timer in ti.reset('time'):
+            >>>     with timer:
+            >>>         kwimage.Affine.affine(offset=offset)
+
         """
-        return cls.affine(offset=offset)
+        offset_ = 0 if offset is None else offset
+        tx, ty = _ensure_iterable2(offset_)
+        # Sympy simplified expression
+        mat = np.array([1.0, 0.0, tx,
+                        0.0, 1.0, ty,
+                        0.0, 0.0, 1.0])
+        mat = mat.reshape(3, 3)  # Faster to make a flat array and reshape
+        self = cls(mat)
+        return self
 
     @classmethod
     def rotate(cls, theta):
@@ -541,6 +584,7 @@ class Affine(Projective):
         return params
 
     @classmethod
+    @profile
     def affine(cls, scale=None, offset=None, theta=None, shear=None,
                about=None, **kwargs):
         """
@@ -572,6 +616,7 @@ class Affine(Projective):
             Affine: the constructed Affine object
 
         Example:
+            >>> from kwimage.transform import *  # NOQA
             >>> rng = kwarray.ensure_rng(None)
             >>> scale = rng.randn(2) * 10
             >>> offset = rng.randn(2) * 10
@@ -642,15 +687,17 @@ class Affine(Projective):
         shear_ = 0 if shear is None else shear
         theta_ = 0 if theta is None else theta
         about_ = 0 if about is None else about
-        sx, sy = _ensure_iterablen(scale_, 2)
-        tx, ty = _ensure_iterablen(offset_, 2)
-        x0, y0 = _ensure_iterablen(about_, 2)
+        sx, sy = _ensure_iterable2(scale_)
+        tx, ty = _ensure_iterable2(offset_)
+        x0, y0 = _ensure_iterable2(about_)
 
         # Make auxially varables to reduce the number of sin/cos calls
+        shear_p_theta = shear_ + theta_
         cos_theta = np.cos(theta_)
         sin_theta = np.sin(theta_)
-        cos_shear_p_theta = np.cos(shear_ + theta_)
-        sin_shear_p_theta = np.sin(shear_ + theta_)
+        cos_shear_p_theta = np.cos(shear_p_theta)
+        sin_shear_p_theta = np.sin(shear_p_theta)
+
         sx_cos_theta = sx * cos_theta
         sx_sin_theta = sx * sin_theta
         sy_sin_shear_p_theta = sy * sin_shear_p_theta
@@ -658,16 +705,35 @@ class Affine(Projective):
         tx_ = tx + x0 - (x0 * sx_cos_theta) + (y0 * sy_sin_shear_p_theta)
         ty_ = ty + y0 - (x0 * sx_sin_theta) - (y0 * sy_cos_shear_p_theta)
         # Sympy simplified expression
-        mat = np.array([[sx_cos_theta, -sy_sin_shear_p_theta, tx_],
-                        [sx_sin_theta,  sy_cos_shear_p_theta, ty_],
-                        [           0,                     0,  1]])
+        mat = np.array([sx_cos_theta, -sy_sin_shear_p_theta, tx_,
+                        sx_sin_theta,  sy_cos_shear_p_theta, ty_,
+                                   0,                     0,  1])
+        mat = mat.reshape(3, 3)  # Faster to make a flat array and reshape
         self = cls(mat)
         return self
 
 
-def _ensure_iterablen(scalar, n):
+# @profile
+# def _ensure_iterablen(scalar, n):
+#     try:
+#         iter(scalar)
+#     except TypeError:
+#         return [scalar] * n
+#     return scalar
+
+
+@profile
+def _ensure_iterable2(scalar):
     try:
-        iter(scalar)
+        a, b = scalar
     except TypeError:
-        return [scalar] * n
-    return scalar
+        a = b = scalar
+    return a, b
+
+if __name__ == '__main__':
+    """
+    CommandLine:
+        python -m kwimage.transform all --profile
+    """
+    import xdoctest
+    xdoctest.doctest_module(__file__)